separator / empty field
- split conf file
- add separator fake field
- filter fields with value
