+++ app/assets/stylesheets/module-investments.scss
@@ -155,6 +155,11 @@
       }
     }
 
+    &--hr {
+      border: 0;
+      border-top: 1px solid $separator;
+    }
+
     &--table-row {
       display: flex;
       align-items: baseline;
