+++ app/javascript/gobierto_investments/webapp/mixins/common.js
@@ -1,184 +1,7 @@
 import { VueFiltersMixin } from "lib/shared";
-import axios from "axios";
 
-const CONFIGURATION = {
-  title: {
-    id: "title_translations"
-  },
-  description: {
-    id: "descripcio-projecte"
-  },
-  phases: {
-    id: "estat"
-  },
-  location: {
-    id: "wkt",
-    center: [41.536908, 2.4418503],
-    minZoom: 13,
-    maxZoom: 16
-  },
-  availableFilters: [
-    {
-      id: "range",
-      startKey: "data-inici",
-      endKey: "data-fin"
-    },
-    {
-      id: "estat",
-      multiple: true
-    },
-    {
-      id: "nom-servei-responsable",
-      multiple: true
-    },
-    {
-      id: "tipus-projecte"
-    },
-    {
-      id: "import"
-    }
-  ],
-  availableGalleryFields: [
-    {
-      id: "data-inici"
-    },
-    {
-      id: "data-final"
-    },
-    {
-      id: "import",
-      filter: "money"
-    },
-    {
-      id: "adjudicatari"
-    }
-  ],
-  availableTableFields: [
-    {
-      id: "nom-projecte"
-    },
-    {
-      id: "estat",
-      multiple: true
-    },
-    {
-      id: "import",
-      filter: "money"
-    }
-  ],
-  availableProjectFields: [
-    {
-      id: "nom-projecte"
-    },
-    {
-      id: "nom-servei-responsable",
-      multiple: true
-    },
-    {
-      id: "tipus-projecte",
-      multiple: true
-    },
-    {
-      id: "descripcio-projecte"
-    },
-    {
-      id: "notes"
-    },
-    {
-      id: "adreca"
-    },
-    {
-      id: "adjudicatari"
-    },
-    {
-      id: "estat",
-      multiple: true
-    },
-    {
-      id: "data-inici",
-      filter: "date"
-    },
-    {
-      id: "data-adjudicacio",
-      filter: "date"
-    },
-    {
-      id: "data-inici-redaccio",
-      filter: "date"
-    },
-    {
-      id: "data-fi-redaccio",
-      filter: "date"
-    },
-    {
-      id: "data-final",
-      type: "icon",
-      icon: {
-        href: "https://twitter.com",
-        name: "file"
-      }
-    },
-    {
-      id: "partida",
-      filter: "money",
-      type: "highlight"
-    },
-    {
-      id: "import",
-      filter: "money"
-    },
-    {
-      id: "import-adjudicacio",
-      filter: "money"
-    },
-    {
-      id: "import-liquidacio",
-      filter: "money"
-    },
-    {
-      id: "tasques",
-      type: "table",
-      table: {
-        columns: ["nomactuacio", "nimport"]
-      }
-    }
-  ],
-  itemSpecialConfiguration: {
-    fn: async data => {
-      const dataArr = Array.isArray(data) ? data : [data];
-      const filterArr = dataArr.filter(d => d.attributes.partida && d.attributes["any-partida"]);
-      const allIds = filterArr.map(d => `'${d.attributes.partida}'`).join(",");
-
-      // Avoid the query if there's no code_with_zone (partida)
-      if (allIds) {
-        const endpoint = `${location.origin}/api/v1/data`;
-        const query = `
-          SELECT code_with_zone as partida, paranyprs as year, sum(parimport) as budget
-          FROM mataro_budgets
-          WHERE "parimport" IS NOT NULL
-          AND "code_with_zone" IN (${allIds})
-          GROUP BY code_with_zone, paranyprs
-        `;
-
-        const { data: { data: sql } } = await axios.get(endpoint, {
-          params: {
-            sql: query.trim()
-          }
-        })
-
-        for (let index = 0; index < dataArr.length; index++) {
-          const { attributes: { partida, "any-partida": year } } = dataArr[index];
-          const { budget } = sql.find(d => d.partida === partida && d.year === year) || {};
-
-          dataArr[index].attributes.old_partida = partida
-          dataArr[index].attributes.partida = budget ? budget : null
-        }
-      }
-
-      return Array.isArray(data) ? dataArr : dataArr[0];
-    }
-  }
-};
+// custom configurations
+import CONFIGURATION from "./mataro.conf.js";
 
 export const baseUrl = `${location.origin}/gobierto_investments/api/v1/projects`;
